requirements.txt
annotated-types==0.6.0
click==8.1.7
greenlet==3.0.1
grpcio==1.59.3
grpcio-tools==1.59.3
joblib==1.3.2
numpy==1.26.2
pandas==2.1.3
protobuf==4.25.1
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
SQLAlchemy==2.0.23
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
